# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Cadastro de Usuários",
    "description": "Documentação das rotas de cadastro e consulta de instituições beneficiárias e pessoas físicas.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/pessoa": {
      "get": {
        "summary": "Consulta de Pessoa Beneficiária",
        "description": "Retorna dados de uma pessoa beneficiária com base no ID informado. Requer autenticação via token.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da pessoa beneficiária"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da pessoa retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "namePer": { "type": "string" },
                    "emailPer": { "type": "string" },
                    "cpfPer": { "type": "string" },
                    "telPer": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "ID não fornecido" },
          "404": { "description": "Pessoa não encontrada" },
          "500": { "description": "Erro no servidor" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/registerPB": {
      "post": {
        "summary": "Cadastro de Pessoa Beneficiária",
        "description": "Registra uma nova pessoa beneficiária no sistema, com verificação de e-mail e CPF.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "namePer": {
                    "type": "string",
                    "example": "João da Silva"
                  },
                  "emailPer": {
                    "type": "string",
                    "format": "email",
                    "example": "joao@email.com"
                  },
                  "passwordPer": {
                    "type": "string",
                    "format": "password",
                    "example": "senha123"
                  },
                  "cpfPer": {
                    "type": "string",
                    "example": "123.456.789-00"
                  },
                  "telPer": {
                    "type": "string",
                    "example": "(11) 91234-5678"
                  }
                },
                "required": ["namePer", "emailPer", "passwordPer", "cpfPer", "telPer"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastro realizado com sucesso. Um e-mail de verificação foi enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cadastro realizado! Verifique seu e-mail."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Campos obrigatórios ausentes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Preencha todos os campos!"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "E-mail ou CPF já cadastrados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "E-mail já cadastrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao verificar ou cadastrar pessoa.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao cadastrar pessoa."
                    }
                  }
                }
              }
            }
          }
        }
      }
      },
    "/verificar-email": {
      "get": {
        "summary": "Verificação de E-mail da Pessoa Beneficiária",
        "description": "Verifica o e-mail de uma pessoa beneficiária através de um token JWT enviado na URL.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
            },
            "description": "Token JWT enviado por e-mail para verificação de conta."
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail verificado com sucesso.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "E-mail verificado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Token ausente na URL.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Token ausente."
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou expirado.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Token inválido ou expirado."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao tentar verificar o e-mail.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Erro interno."
                }
              }
            }
          }
        }
      }
    },
...
     "/cadastrar-itemP": {
      "post": {
        "summary": "Cadastrar Pedido por Pessoa Beneficiária",
        "description": "Cadastra um novo pedido feito por uma pessoa beneficiária. Todos os campos são obrigatórios.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name_item",
                  "description",
                  "quantity_item",
                  "category",
                  "urgencia_enum",
                  "locate"
                ],
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 7,
                    "description": "ID da pessoa beneficiária"
                  },
                  "name_item": {
                    "type": "string",
                    "example": "Cesta Básica",
                    "description": "Nome do item solicitado"
                  },
                  "description": {
                    "type": "string",
                    "example": "Itens essenciais para alimentação de família com 4 pessoas.",
                    "description": "Descrição do item"
                  },
                  "quantity_item": {
                    "type": "integer",
                    "example": 3,
                    "description": "Quantidade desejada"
                  },
                  "category": {
                    "type": "string",
                    "example": "Alimentos",
                    "description": "Categoria do item"
                  },
                  "urgencia_enum": {
                    "type": "string",
                    "enum": ["Urgente", "Alta", "Média", "Baixa"],
                    "example": "Urgente",
                    "description": "Nível de urgência"
                  },
                  "locate": {
                    "type": "string",
                    "example": "Rua das Acácias, 123 - SP",
                    "description": "Localização para entrega"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido cadastrado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Pedido realizado com sucesso."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida (campos obrigatórios ausentes ou ID inválido).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Preencha todos os campos!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao cadastrar o pedido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno ao criar o pedido."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/pedidosP": {
      "get": {
        "summary": "Listar pedidos por pessoa beneficiária",
        "description": "Retorna todos os pedidos feitos por uma pessoa beneficiária específica com base no ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "ID da pessoa beneficiária",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "pessoa_beneficiaria_id": { "type": "integer" },
                      "name_item": { "type": "string" },
                      "description": { "type": "string" },
                      "quantity_item": { "type": "integer" },
                      "category": { "type": "string" },
                      "urgencia_enum": { "type": "string" },
                      "locate": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
...
"/deletepessoa": {
  "delete": {
    "summary": "Deletar pessoa beneficiária",
    "description": "Remove uma pessoa beneficiária e seus pedidos associados do banco de dados. Requer autenticação com token JWT.",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "parameters": [
      {
        "name": "id",
        "in": "query",
        "required": true,
        "description": "ID da pessoa beneficiária a ser deletada",
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Usuário deletado com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Usuário deletado com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "ID da pessoa não fornecido"
      },
      "404": {
        "description": "Usuário não encontrado"
      },
      "500": {
        "description": "Erro ao excluir pedidos ou usuário"
      }
    }
  }
},
...
"/pessoaup/{id}": {
  "put": {
    "summary": "Atualizar dados da pessoa beneficiária",
    "description": "Atualiza nome, email e telefone da pessoa beneficiária. Exige que a senha atual seja informada corretamente.",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID da pessoa beneficiária a ser atualizada",
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "required": ["namePer", "emailPer", "passwordPer", "telPer"],
            "properties": {
              "namePer": {
                "type": "string",
                "description": "Nome da pessoa beneficiária"
              },
              "emailPer": {
                "type": "string",
                "format": "email",
                "description": "Email da pessoa beneficiária"
              },
              "passwordPer": {
                "type": "string",
                "description": "Senha atual para validação"
              },
              "telPer": {
                "type": "string",
                "description": "Telefone da pessoa beneficiária"
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Pessoa atualizada com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Pessoa atualizada com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Campos obrigatórios não preenchidos"
      },
      "403": {
        "description": "Senha incorreta"
      },
      "404": {
        "description": "Pessoa não encontrada"
      },
      "500": {
        "description": "Erro no servidor"
      }
    }
  }
},
...
}
}